ENGINES/requirements.txt
# Core dependencies for WitnessOS Divination Engines
# Phase 1: Foundation requirements

# Data validation and models
pydantic>=2.0.0

# Date/time handling
pytz>=2023.3

# Testing framework
pytest>=7.0.0

# Phase 2+ dependencies (will be uncommented as needed)

# Astronomical calculations (Phase 3)
pyswisseph>=2.10.0

# Mathematical and visualization
# numpy>=1.24.0
# matplotlib>=3.7.0

# Image generation and graphics
# pillow>=10.0.0
# cairo-python>=1.20.0

# Optional advanced features
# svglib>=1.5.0
# reportlab>=4.0.0

# Development dependencies
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
